Added More Powers
Not all added... still debating whther or not I should complex this. Should I be give wither to the player, or surrounding players... how does this connect to ur friend... Maybe all players in the radius of both u and the friend, and once one takes it the other can/t?